refactor(optimiser): restructure module organization and enhance CORS handling
